pyproject.toml
[project]
name = "summarize-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.0",
    "fastapi>=0.123.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-pymupdf4llm>=0.5.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
